+++ src/main/java/racingcar/Input.java
@@ -1,4 +1,4 @@
-package step3;
+package racingcar;
 
 import java.util.Scanner;
