+++ src/main/java/racingcar/numbergenerator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package step3.numbergenerator;
+package racingcar.numbergenerator;
 
 public interface NumberGenerator {
