+++ src/main/java/racingcar/numbergenerator/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package step3.numbergenerator;
+package racingcar.numbergenerator;
 
 import java.util.Random;
