+++ src/test/java/positiveinteger/StringCalculatorTest.java
@@ -1,4 +1,4 @@
-package step2;
+package positiveinteger;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
