+++ src/test/java/racingcar/RaceTest.java
@@ -1,10 +1,8 @@
-package step3;
+package racingcar;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class RaceTest {
